# Benchmark and improve ResourceIdentifier#equals