## Before this PR
ResourceIdentifier#equals did not compare hashCode()

## After this PR
The `com.palantir.ri.ResourceIdentifier#equals(Object)` method is important for performance as many services rely on comparing equality of RIDs. We can improve the performance of mismatches by utilizing the memoized hashCode when comparing two non-identity-equal RIDs before doing the byte-wise/character String equals comparison.

==COMMIT_MSG==
ResourceIdentifier#equals compares hashCode() before RID string comparison
==COMMIT_MSG==

Before
======
```
Apple M1 Pro, JDK 15.0.7, OpenJDK 64-Bit Server VM, 15.0.7+4-MTS
Benchmark                                                         Mode  Cnt   Score    Error   Units
ResourceIdentifierHashEqualsBenchmark.equals                      avgt    3   2.797 ±  0.099   ns/op
ResourceIdentifierHashEqualsBenchmark.equals:·gc.alloc.rate       avgt    3   0.010 ±  0.279  MB/sec
ResourceIdentifierHashEqualsBenchmark.equals:·gc.alloc.rate.norm  avgt    3  ≈ 10⁻⁵             B/op
ResourceIdentifierHashEqualsBenchmark.equals:·gc.count            avgt    3     ≈ 0           counts
ResourceIdentifierHashEqualsBenchmark.hash                        avgt    3   0.784 ±  0.118   ns/op
ResourceIdentifierHashEqualsBenchmark.hash:·gc.alloc.rate         avgt    3   0.006 ±  0.140  MB/sec
ResourceIdentifierHashEqualsBenchmark.hash:·gc.alloc.rate.norm    avgt    3  ≈ 10⁻⁶             B/op
ResourceIdentifierHashEqualsBenchmark.hash:·gc.count              avgt    3     ≈ 0           counts
```
```
Apple M1 Pro, JDK 17.0.3, OpenJDK 64-Bit Server VM, 17.0.3+7-LTS
Benchmark                                                         Mode  Cnt   Score    Error   Units
ResourceIdentifierHashEqualsBenchmark.equals                      avgt    3   2.126 ±  2.926   ns/op
ResourceIdentifierHashEqualsBenchmark.equals:·gc.alloc.rate       avgt    3   0.003 ±  0.071  MB/sec
ResourceIdentifierHashEqualsBenchmark.equals:·gc.alloc.rate.norm  avgt    3  ≈ 10⁻⁶             B/op
ResourceIdentifierHashEqualsBenchmark.equals:·gc.count            avgt    3     ≈ 0           counts
ResourceIdentifierHashEqualsBenchmark.hash                        avgt    3   0.359 ±  0.026   ns/op
ResourceIdentifierHashEqualsBenchmark.hash:·gc.alloc.rate         avgt    3   0.003 ±  0.070  MB/sec
ResourceIdentifierHashEqualsBenchmark.hash:·gc.alloc.rate.norm    avgt    3  ≈ 10⁻⁶             B/op
ResourceIdentifierHashEqualsBenchmark.hash:·gc.count              avgt    3     ≈ 0           counts
```

After
=====
```
Apple M1 Pro, JDK 15.0.7, OpenJDK 64-Bit Server VM, 15.0.7+4-MTS
Benchmark                                                         Mode  Cnt   Score    Error   Units
ResourceIdentifierHashEqualsBenchmark.equals                      avgt    3   1.812 ±  3.899   ns/op
ResourceIdentifierHashEqualsBenchmark.equals:·gc.alloc.rate       avgt    3   0.004 ±  0.072  MB/sec
ResourceIdentifierHashEqualsBenchmark.equals:·gc.alloc.rate.norm  avgt    3  ≈ 10⁻⁶             B/op
ResourceIdentifierHashEqualsBenchmark.equals:·gc.count            avgt    3     ≈ 0           counts
ResourceIdentifierHashEqualsBenchmark.hash                        avgt    3   0.791 ±  0.090   ns/op
ResourceIdentifierHashEqualsBenchmark.hash:·gc.alloc.rate         avgt    3   0.004 ±  0.074  MB/sec
ResourceIdentifierHashEqualsBenchmark.hash:·gc.alloc.rate.norm    avgt    3  ≈ 10⁻⁶             B/op
ResourceIdentifierHashEqualsBenchmark.hash:·gc.count              avgt    3     ≈ 0           counts
```
```
Apple M1 Pro, JDK 17.0.3, OpenJDK 64-Bit Server VM, 17.0.3+7-LTS
Benchmark                                                         Mode  Cnt   Score    Error   Units
ResourceIdentifierHashEqualsBenchmark.equals                      avgt    3   1.180 ±  0.086   ns/op
ResourceIdentifierHashEqualsBenchmark.equals:·gc.alloc.rate       avgt    3   0.004 ±  0.072  MB/sec
ResourceIdentifierHashEqualsBenchmark.equals:·gc.alloc.rate.norm  avgt    3  ≈ 10⁻⁶             B/op
ResourceIdentifierHashEqualsBenchmark.equals:·gc.count            avgt    3     ≈ 0           counts
ResourceIdentifierHashEqualsBenchmark.hash                        avgt    3   0.355 ±  0.047   ns/op
ResourceIdentifierHashEqualsBenchmark.hash:·gc.alloc.rate         avgt    3   0.004 ±  0.072  MB/sec
ResourceIdentifierHashEqualsBenchmark.hash:·gc.alloc.rate.norm    avgt    3  ≈ 10⁻⁶             B/op
ResourceIdentifierHashEqualsBenchmark.hash:·gc.count              avgt    3     ≈ 0           counts
```

## Possible downsides?
This adds an int comparison to hot equals code path, and in cases where `ResourceIdentifier` is a key for a hash structure the `hashCode()` may have just been compared and found to be equal, but in non-hash-based structures such as trees and graphs including `hashCode()` comparison in `equals()`