@svc-autorelease 2022-07-12
Released [2.5.0](https://github.com/palantir/resource-identifier/releases/tag/2.5.0)